backend/requirements.txt
# Web framework and server
fastapi>=0.100.0,<0.112.0
uvicorn[standard]>=0.20.0,<0.30.0
pydantic>=2.4.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0  # JWT handling
passlib[bcrypt]>=1.7.4,<2.0.0  # Password hashing

# Database and ORM
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.12.0,<2.0.0
psycopg2-binary>=2.9.9,<3.0.0  # PostgreSQL driver
python-dotenv>=1.0.0,<2.0.0   # Environment variables

# Numerical and scientific computing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scipy>=1.10.0,<2.0.0
statsmodels>=0.14.0,<0.15.0

# Time series processing
pmdarima>=2.0.3,<3.0.0  # ARIMA models
openpyxl>=3.1.2,<4.0.0  # Excel file support

# Testing and development
pytest>=7.3.1,<8.0.0
httpx>=0.24.1,<0.25.0  # HTTP client for testing
